Refactoring crypto_pufs.c to retain pufcc.c in a modular way
